+++ src/index.ts
@@ -17,6 +17,6 @@
 export * from "./types.js";
 export * from "./client.js";
 export * from "./verify.js";
-export { canonicalize, verifySignature } from "./crypto.js";
+export { canonicalize, verifySignature, verifyWebhookSignature } from "./crypto.js";
 export * from "./integrations/index.js";
 export { signPayload } from "./sign.js";